+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.2.85 (unreleased)
 -------------------
 
+- WEB-4153 : Add a refresh_modification_date view to refresh the modification date of SectionEvents and SectionNews
+  [remdub]
+
 - WEB-4153 : Add a cache key on SectionEvents and SectionNews requests to refresh the cache when the section is modified
   [remdub]